+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keep-changelog]
 and this project adheres to [Semantic Versioning][semver].
 
-## [0.0.1] - 2021-04-14
+## [1.0.0] - 2021-05-21
 
 ### Added
 
-- TMS-31: Add materials block
-- TMS-60: Register post type material and taxonomy material type
+- TMS-32: Add file layout for Accordion block #6
+- TMS-31: Add materials block #5
+- TMS-60: Register post type material and taxonomy material type #4
 
 
 [keep-changelog]: http://keepachangelog.com/en/1.0.0/